Update SigmoidFocalLoss_cuda.cu
Converted to long to support windows10

+++ maskrcnn_benchmark/csrc/cuda/SigmoidFocalLoss_cuda.cu
@@ -162,7 +162,7 @@ at::Tensor SigmoidFocalLoss_backward_cuda(
   auto d_logits_size = num_samples * logits.size(1);
   cudaStream_t stream = at::cuda::getCurrentCUDAStream();
 
-  dim3 grid(std::min(THCCeilDiv(d_logits_size, 512L), 4096L));
+  dim3 grid(std::min(THCCeilDiv((long)d_logits_size, 512L), 4096L));
   dim3 block(512);
 
   if (d_logits.numel() == 0) {
